+++ src/Transit.cpp
@@ -742,6 +742,20 @@ struct TransitModule : TransitBase<NUM_PRESETS> {
 		presetClear(p);
 	}
 
+	void presetShiftFront(int p) {
+		for (int i = 1; i <= p; i++) {
+			TransitSlot* slot = expSlot(i);
+			if (*(slot->presetSlotUsed)) {
+				presetCopyPaste(i, i - 1);
+				*expSlotLabel(i - 1) = *expSlotLabel(i);
+			}
+			else {
+				presetClear(i - 1);
+			}
+		}
+		presetClear(p);
+	}
+
 	void setProcessDivision(int d) {
 		presetProcessDivision = d;
 		presetProcessDivider.setDivision(presetProcessDivision);
@@ -786,8 +800,11 @@ struct TransitModule : TransitBase<NUM_PRESETS> {
 			case SLOT_CMD::SAVE:
 				presetSave(i);
 				return -1;
-			case SLOT_CMD::SHIFTBACK:
+			case SLOT_CMD::SHIFT_BACK:
 				presetShiftBack(i);
+				return -1;
+			case SLOT_CMD::SHIFT_FRONT:
+				presetShiftFront(i);
 				return -1;
 			default:
 				return -1;