+++ CHANGELOG.md
@@ -46,7 +46,7 @@
 - Module [TRANSIT](./docs/Transit.md)
     - Added context menu option for unbinding all bound parameters of a module (#268)
     - Added "Auto"-mode besides "Read" and "Write" ([manual](./docs/Transit.md#auto-mode)) (#269)
-    - Added "Shift back" context menu option (#274)
+    - Added "Shift front" and "Shift back" context menu options (#274)
 
 ## 1.9.0